+++ packages/cfg/lib/Pair.ts
@@ -1,0 +1,4 @@
+export interface Pair<T, U> {
+  first: T;
+  second: U;
+}