+++ packages/core/test/mocks/DummyCFG.mock.ts
@@ -1,0 +1,3 @@
+import { CFG } from "@syntest/cfg";
+
+export class DummyCFG extends CFG {}