+++ packages/plugin-core-example/test/ExamplePlugin.test.ts
@@ -1,0 +1,33 @@
+import { DummyLauncher } from "./mocks/DummyLauncher.mock";
+import { CommandLineInterface, EventManager } from "@syntest/core";
+import { DummyEncoding } from "./mocks/DummyEncoding.mock";
+
+describe("", () => {
+  it("Valid path", () => {
+    const eventManager = new EventManager({});
+    const cli = new CommandLineInterface();
+    const launcher = new DummyLauncher<DummyEncoding>(
+      "dummy",
+      eventManager,
+      cli
+    );
+    const args = [
+      "--use_plugin",
+      "../packages/plugin-core-example/lib",
+      "--target_root_directory",
+      ".",
+    ];
+    launcher.run(args);
+  });
+  it("Invalid path", () => {
+    const eventManager = new EventManager({});
+    const cli = new CommandLineInterface();
+    const launcher = new DummyLauncher<DummyEncoding>(
+      "dummy",
+      eventManager,
+      cli
+    );
+    const args = ["--use_plugin", "../plugins", "--target_root_directory", "."];
+    launcher.run(args);
+  });
+});