+++ test/mocks/DummyCFG.mock.ts
@@ -1,3 +1,0 @@
-import { CFG } from "../../src";
-
-export class DummyCFG extends CFG {}